Refactor tscn lexer and parser (#237)
In the initial implementation, lexing and parsing were conflated. Supporting any new constructs was lots of work because we needed to be explicit about each token.

This refactors that part to be more in line with good config parsing practices.
Now, errors are reported with miette in a fancy way with link to the LoC of the .tscn file that broke and a reason why that line is not adequate.

There are still some code health issues in the part where we spawn the tree into bevy world but the rscn code itself looks pretty good.